refactor: 旅行カードの分離
旅行リスト一覧タブのファイルから旅行カードの関数を別ファイルに分離しました。
